pyproject.toml
[tool.poetry]
name = "gcp-pilot"
version = "1.14.0"
description = "Google Cloud Platform Friendly Pilot"
authors = ["Joao Daher <joao@daher.dev>"]
repository = "https://github.com/flamingo-run/gcp-pilot"
readme = "README.md"
keywords = ["gcp", "google cloud python", "datastore orm"]
packages = [
    { include = "gcp_pilot" },
]


[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pytz = "*"
tenacity = "*"
factory_boy = "*"
google-api-python-client = "*"
google-cloud-tasks = { version = "^2", optional = true}
google-cloud-scheduler = { version = "^2", optional = true}
google-cloud-storage = { version = "^2", optional = true}
google-cloud-build = { version = "^3", optional = true}
google-cloud-bigquery = {version = "^3", optional = true, extras = ["storage"]}
google-cloud-speech = { version = "^2", optional = true}
gspread = { version = "^5", optional = true}
google-cloud-pubsub = { version = "^2", optional = true}
google-cloud-datastore = { version = "^2", optional = true}
google-cloud-iam = { version = "*", optional = true}
pydantic = { version = "*", optional = true}
google-cloud-dns = { version = "^0", optional = true}
google-cloud-logging = { version = "^3", optional = true}
google-cloud-error-reporting = { version = "*", optional = true}
google-cloud-secret-manager = { version = "*", optional = true}
"fhir.resources" = { version = "*", optional = true}

[tool.poetry.extras]
tasks = ["google-cloud-tasks", "google-cloud-scheduler"]
build = ["google-cloud-build"]
storage = ["google-cloud-storage"]
bigquery = ["google-cloud-bigquery", "google-cloud-storage"]
speech = ["google-cloud-speech"]
sheets = ["gspread"]
pubsub = ["google-cloud-pubsub"]
datastore = ["google-cloud-datastore", "pydantic"]
dns = ["google-cloud-dns"]
secret = ["google-cloud-secret-manager"]
monitoring = ["google-cloud-logging", "google-cloud-error-reporting"]
healthcare = ["fhir.resources"]
iam = ["google-cloud-iam"]


[tool.poetry.dev-dependencies]
black = "*"
ruff = "*"
pyjwt = "*"
pytest-cov = "*"
pytest-rerunfailures = "*"
pytest-xdist = "*"
pytest-sugar = "*"
pytest-random-order = "*"
pytest-timeout = "*"

[tool.black]
line-length = 120
target-version = ['py311']


[tool.ruff]
select = ["RUF", "I", "PL", "F", "COM", "UP", "DJ", "T10", "T20", "DTZ", "SIM", "TID", "PTH", "ERA", "TRY"]
ignore = ["PLR2004", "PLR0911", "PLR0912", "PLR0913", "PLR0915", "DJ001", "DJ001", "DJ008", "TRY003"]
line-length = 120
target-version = "py311"


[tool.pytest.ini_options]
addopts = "-n auto --reruns 2 --random-order --timeout 100 --cov"
python_files = ["*_test.py"]


[tool.coverage.run]
branch = true
source = ["gcp_pilot"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
